move shared module and its declared elements to shared folder
