+++ src/app/modules/layout/layout.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { MainComponent } from './main/main.component';
 import { HeaderComponent } from './header/header.component';
-import { SharedModule } from '../shared.module';
+import { SharedModule } from '../shared/shared.module';
 import { OplContainerComponent } from './opl-container/opl-container.component';
 import { UserStatusComponent } from './header/user-status/user-status.component';
 import { RouterModule } from '@angular/router';
