+++ src/app/modules/shared/avatar/avatar.component.scss
