+++ src/app/modules/shared/avatar/avatar.component.spec.ts
