+++ src/app/modules/shared/avatar/avatar.component.ts
