+++ src/app/modules/shared/resizable.directive.ts
